+++ CHANGES.md
@@ -1,10 +1,12 @@
-CHANGES - 2.3rc1 - 2019-05-13
+CHANGES - 2.3rc1 - 2019-05-15
 =============================
 
 
 Changes in CUPS v2.3rc1
 -----------------------
 
+- The `cupsctl` command now prevents setting "cups-files.conf" directives
+  (Issue #5530)
 - The footer in the web interface covered some content on small displays
   (Issue #5574)
 - The `ippeveprinter` command now looks for print commands in the "command"